backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
requests==2.31.0
slowapi==0.1.9
limits==3.10.0
python-dotenv==1.0.1
sqlite3
Pillow==10.2.0
opencv-python==4.9.0.80
pydantic==2.6.1
sse-starlette==2.0.0
pandas==2.2.0

